Move all components into 'components' directory
