+++ client/app/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './login/login.component';
+import { HomeComponent } from './components/home/home.component';
+import { LoginComponent } from './components/login/login.component';
 import { Guard } from './guard';
-import { VolunteerCheckInComponent } from './volunteer-check-in/volunteer-check-in.component';
+import { VolunteerCheckInComponent } from './components/volunteer-check-in/volunteer-check-in.component';
 
 const routes: Routes = [
   { path: 'login', component: LoginComponent },
