+++ client/app/app.module.ts
@@ -19,19 +19,19 @@ import 'hammerjs';
 import { AppRoutingModule } from './app-routing.module';
 
 import { AppComponent } from './app.component';
-import { LayoutComponent } from './shared/layout/layout.component';
-import { HeaderComponent } from './shared/layout/header/header.component';
-import { FooterComponent } from './shared/layout/footer/footer.component';
-import { LoginComponent } from './login/login.component';
-import { HomeComponent } from './home/home.component';
-import { VolunteerMenuComponent } from './home/volunteer-menu/volunteer-menu.component';
-import { VolunteerCheckInComponent } from './volunteer-check-in/volunteer-check-in.component';
-import { CheckInFormComponent } from './volunteer-check-in/check-in-form/check-in-form.component';
-import { SignatureFieldComponent } from './signature-field/signature-field.component';
-import { NewVolunteerFormComponent } from './volunteer-check-in/new-volunteer-form/new-volunteer-form.component';
-import { VisitDataDisplayComponent } from './visit-data-display/visit-data-display.component';
-import { VisitHistoryTableComponent } from './visit-data-display/visit-history-table/visit-history-table.component';
-import { DailyHoursChartComponent } from './visit-data-display/daily-hours-chart/daily-hours-chart.component';
+import { LayoutComponent } from './components/shared/layout/layout.component';
+import { HeaderComponent } from './components/shared/layout/header/header.component';
+import { FooterComponent } from './components/shared/layout/footer/footer.component';
+import { LoginComponent } from './components/login/login.component';
+import { HomeComponent } from './components/home/home.component';
+import { VolunteerMenuComponent } from './components/home/volunteer-menu/volunteer-menu.component';
+import { VolunteerCheckInComponent } from './components/volunteer-check-in/volunteer-check-in.component';
+import { CheckInFormComponent } from './components/volunteer-check-in/check-in-form/check-in-form.component';
+import { SignatureFieldComponent } from './components/volunteer-check-in/check-in-form/signature-field/signature-field.component';
+import { NewVolunteerFormComponent } from './components/volunteer-check-in/new-volunteer-form/new-volunteer-form.component';
+import { VisitDataDisplayComponent } from './components/visit-data-display/visit-data-display.component';
+import { VisitHistoryTableComponent } from './components/visit-data-display/visit-history-table/visit-history-table.component';
+import { DailyHoursChartComponent } from './components/visit-data-display/daily-hours-chart/daily-hours-chart.component';
 
 import { UserService } from './services/user.service';
 import { VolunteerService } from './services/volunteer.service';
