+++ client/app/components/home/volunteer-menu/volunteer-menu.component.spec.ts
@@ -2,7 +2,7 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { VolunteerMenuComponent } from './volunteer-menu.component';
 import { MdCardModule } from '@angular/material';
-import { MockVisitService, VisitService } from '../../services/visit.service';
+import { MockVisitService, VisitService } from '../../../services/visit.service';
 
 describe('VolunteerMenuComponent', () => {
   let component: VolunteerMenuComponent;
