+++ client/app/components/login/login.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms'
 import { Router } from '@angular/router';
-import { UserService } from '../services/user.service';
-import { VisitService } from '../services/visit.service';
+import { UserService } from '../../services/user.service';
+import { VisitService } from '../../services/visit.service';
 
 @Component({
   selector: 'app-login',
