+++ client/app/components/shared/layout/footer/footer.component.ts
