+++ client/app/components/shared/layout/header/header.component.ts
