+++ client/app/components/shared/layout/layout.component.spec.ts
