+++ client/app/components/shared/layout/layout.component.ts
