+++ client/app/components/visit-data-display/daily-hours-chart/daily-hours-chart.component.html
