+++ client/app/components/visit-data-display/daily-hours-chart/daily-hours-chart.component.spec.ts
@@ -3,9 +3,8 @@ import { StoreModule } from '@ngrx/store';
 import { ChartsModule } from 'ng2-charts';
 
 import { DailyHoursChartComponent } from './daily-hours-chart.component';
-import { testVisits } from '../../models/visit';
-import { visitReducer } from '../../reducers/visit';
-import { forEach } from '@angular/router/src/utils/collection';
+import { testVisits } from '../../../models/visit';
+import { visitReducer } from '../../../reducers/visit';
 
 describe('DailyHoursChartComponent', () => {
   let component: DailyHoursChartComponent;
