+++ client/app/components/visit-data-display/daily-hours-chart/daily-hours-chart.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Visit } from '../../models/visit';
+import { Visit } from '../../../models/visit';
 
 @Component({
   selector: 'app-daily-hours-chart',
