+++ client/app/components/volunteer-check-in/check-in-form/signature-field/signature-field.component.ts
