+++ client/app/components/volunteer-check-in/new-volunteer-form/new-volunteer-form.component.spec.ts
@@ -2,11 +2,11 @@ import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { AbstractControl, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MdAutocompleteModule, MdInputModule, MdSnackBarModule } from '@angular/material';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { StoreModule } from '@ngrx/store';
 
 import { NewVolunteerFormComponent } from './new-volunteer-form.component';
-import { MockVolunteerService, VolunteerService } from '../../services/volunteer.service';
-import { StoreModule } from '@ngrx/store';
-import { volunteerReducer } from '../../reducers/volunteer';
+import { MockVolunteerService, VolunteerService } from '../../../services/volunteer.service';
+import { volunteerReducer } from '../../../reducers/volunteer';
 
 describe('NewVolunteerFormComponent', () => {
   let component: NewVolunteerFormComponent,
